Provisioning: Move repository limits tests into quota package
Move the integration tests from repository/limits into the quota
package so they share the same Grafana server and use dynamic quota
configuration via SetQuotaStatus instead of requiring a dedicated
server startup with custom GrafanaOpts.

This eliminates the limits package entirely — one fewer server to
start in CI.

Made-with: Cursor
